+++ src/oDesk/API/constants.php
@@ -10,6 +10,8 @@
  * @license     oDesk's API Terms of Use {@link http://developers.odesk.com/API-Terms-of-Use}
  */
 
+define('ODESK_API_EP_NAME', 'api');
+define('ODESK_GDS_EP_NAME', 'gds');
 define('ODESK_BASE_URL', 'https://www.odesk.com');
-define('ODESK_BASE_URL_API', ODESK_BASE_URL . '/api');
-define('ODESK_BASE_URL_GDS', ODESK_BASE_URL . '/gds');
+define('ODESK_BASE_URL_API', ODESK_BASE_URL . '/' . ODESK_API_EP_NAME);
+define('ODESK_BASE_URL_GDS', ODESK_BASE_URL . '/' . ODESK_GDS_EP_NAME);